Staging 4.10.2

+++ teamcity/PublishCSPForUnity.py
@@ -230,7 +230,8 @@ def main():
         f.write(minimal_meta_template.render(guid=generate_guid(), importer_type="TextScript"))
 
     # Copy readme
-    shutil.copy(f"{config.default_output_directory}/README.md", args.output_directory)
+    if (os.path.exists(f"{config.default_output_directory}/README.md")):
+        shutil.copy(f"{config.default_output_directory}/README.md", args.output_directory)
 
     # Create and publish NPM package
     os.chdir(args.output_directory)
